chore: rename widelogo to use kebab-case
