+++ app/components/ui/wide-logo.tsx
